rename google maps doctype, add helper methods
