+++ frappe/integrations/doctype/google_maps_settings/google_maps_settings.js
@@ -1,5 +1,5 @@
 // Copyright (c) 2017, Frappe Technologies and contributors
 // For license information, please see license.txt
 
-frappe.ui.form.on('Google Maps', {
+frappe.ui.form.on('Google Maps Settings', {
 });
