+++ frappe/integrations/doctype/google_maps_settings/test_google_maps_settings.js
@@ -2,7 +2,7 @@
 // rename this file from _test_[name] to test_[name] to activate
 // and remove above this line
 
-QUnit.test("test: Google Maps", function (assert) {
+QUnit.test("test: Google Maps Settings", function (assert) {
 	let done = assert.async();
 
 	// number of asserts
@@ -10,7 +10,7 @@ QUnit.test("test: Google Maps", function (assert) {
 
 	frappe.run_serially([
 		// insert a new Google Maps
-		() => frappe.tests.make('Google Maps', [
+		() => frappe.tests.make('Google Maps Settings', [
 			// values to be set
 			{key: 'value'}
 		]),
