+++ frappe/patches/v11_0/rename_google_maps_doctype.py
@@ -1,0 +1,7 @@
+import frappe
+from frappe.model.rename_doc import rename_doc
+
+def execute():
+	if frappe.db.table_exists("Google Maps") and not frappe.db.table_exists("Google Maps Settings"):
+		rename_doc('DocType', 'Google Maps', 'Google Maps Settings')
+		frappe.reload_doc('integrations', 'doctype', 'google_maps_settings')
